feat(orga): moved usecasefile from lib to prescription
